WIP: add gbc support
