+++ PixelEmu/views/save_states/SaveStateEntriesView.swift
@@ -41,8 +41,8 @@ struct SaveStateEntriesView: View {
         // create a new save state
 
         if let emu = emulator {
-            let dataPtr = emu.createSaveState()
-            let compressedLength = emu.compressedLength()
+            let dataPtr = try! emu.createSaveState()
+            let compressedLength = try! emu.compressedLength()
 
             let unsafeBufferPtr = UnsafeBufferPointer(start: dataPtr, count: Int(compressedLength))
